test: properly document the #expect_execution feature
